minor communication bug fixed

+++ src/main/java/at/asdf00/avaplus/objects/blocks/amogus/ContainerAmogus.java
@@ -48,7 +48,7 @@ public class ContainerAmogus extends Container {
     @Override
     public void updateProgressBar(int id, int data) {
         if (id == 0)
-            tileentity.setRfConsumedScaled(data);
+            tileentity.setRfConsumedScaled((short)data);
     }
     @Override
     public void detectAndSendChanges() {
